Set PWM triggers using udev rules & fix PWM DT
Let's keep DT clean and do not assign triggers in device tree. The
functions of PWM LEDs (D1..D4) are now set by udev rules and can be
changed by the user via sysfs.

LEDs on Unleashed are wired to supply instead of ground thus we need
to set "active-low".

Signed-off-by: David Abdurachmanov <david.abdurachmanov@sifive.com>