linux-mainline_5.12: rebase to v5.12.18
Rebase existing v5.12 kernel tree to v5.12.18. The patch to fix cache
controller interrupts was already merged.

Remove cfg/fs/vfat.scc from KERNEL_FEATURES as this is not part of oe-core
layer. This is a recent addition to linux-yocto.inc. The previous *.scc are
already removed by meta-riscv layer.

Signed-off-by: David Abdurachmanov <david.abdurachmanov@sifive.com>