Sync with upstream layers (new override syntax)
This is a large update mostly due to a new override syntax. The old "_"
character was replaced with ":".

This layer is now only compatible with honister release.

There are two packages updated:
- xf86-video-amdgpu was updated from 19.1.0 to 21.0.0
- mesa was updated from 21.1.2 to 21.1.5

Two packages were added to Xfce4 desktop disk image:
- xf86-video-modesetting
- xf86-video-fbdev

The root password in the recipe is now SHA512 encrypted.

Signed-off-by: David Abdurachmanov <david.abdurachmanov@sifive.com>