Update stress-ng to 0.13.05 and fix source URL
The developer of stress-ng has left Canonical, and the project
has moved to github. However, the releases there are not
drop in replacements for the old release tarballs. So I
decided to pull a patch release newer version from the
debian archive of the original source instead to get the
build working again. It's a bit of a hack, but I'm not familiar
enough with bitbake to make a tagged pull from github work
correctly after a few attempts.

For reference, the new home of the project is:
https://github.com/ColinIanKing/stress-ng